refactor: lift shared sheet actions and tab wiring into the base
効果・装備のトグル、技能の作成と段入力、カスタム技能の能力値selectの受け口は
共鳴者専用の作りではなく、getEmbeddedDocument とアイテムの型述語だけで動く。
NPC・怪異シートがタブと技能行を持てるよう EmokloreActorSheet へ移す。
タブナビのパートと context.tab の代入も同じ理由で基底に置く（タブを持たない
シートでは tabs が無いので何もしない）。
