+++ module/applications/actor-sheet.ts
@@ -1,4 +1,4 @@
-import type { ApplicationRenderContext } from "@client/applications/_types.mjs";
+import type { ApplicationRenderContext, ApplicationTab } from "@client/applications/_types.mjs";
 import { isBaseSkillKey } from "../config/base-skills";
 import { isSkillKey } from "../config/skills";
 import type { SkillRef } from "../data/character-like";
@@ -9,7 +9,9 @@ import {
   getEmbeddedDocument,
   resolveEmbeddedDocumentClass,
 } from "../utils/sheet";
+import { promptCreateSkill } from "./dialogs/create-skill-dialog";
 import EmokloreDocumentSheetMixin from "./document-sheet-mixin";
+import { resolveSegmentValue } from "./helpers";
 import { requestResonanceRoll, requestSkillRoll } from "./rolls";
 import type {
   EmokloreActorSheetOptions,
@@ -43,8 +45,17 @@ export class EmokloreActorSheet extends EmokloreDocumentSheetMixin(
       viewDoc: this._viewDoc,
       createDoc: this._createDoc,
       deleteDoc: this._deleteDoc,
+      toggleEffect: this._toggleEffect,
+      toggleEquipped: this._toggleEquipped,
+      selectSegment: this._selectSegment,
+      createSkill: this._createSkill,
     },
   };
+
+  // タブナビのパート。本体のテンプレートなので systemPath を通さない。
+  // 各シートの PARTS はクラスごとに丸ごと宣言し直す決まり（HandlebarsApplicationMixin は
+  // PARTS を継承マージしない）ので、パスをここに置いて写し間違いを防ぐ
+  static readonly TAB_NAV_PART = { template: "templates/generic/tab-navigation.hbs" };
 
   /**
    * 行の埋め込みドキュメントを開く・作る・消す。
@@ -69,6 +80,144 @@ export class EmokloreActorSheet extends EmokloreDocumentSheetMixin(
     const docData = createDocumentData(target, this.actor);
     const docCls = resolveEmbeddedDocumentClass(target.dataset.documentClass);
     await docCls.create(docData, { parent: this.actor });
+  }
+
+  static async _toggleEffect(this: EmokloreActorSheet, _event: Event, target: HTMLElement) {
+    const effect = getEmbeddedDocument(target, this.actor);
+    if (effect) await effect.update({ disabled: !effect.disabled });
+  }
+
+  /**
+   * アイテムタブの装備チェックボックス。
+   *
+   * アイテムの値の編集だがシートのフォームには載せられない（同じ name の入力を
+   * 2箇所に描けない）ので、name を持たないチェックボックスから直接アイテムへ書く。
+   */
+  static async _toggleEquipped(this: EmokloreActorSheet, _event: Event, target: HTMLElement) {
+    const item = getEmbeddedDocument(target, this.actor);
+    if (item) await item.update({ "system.equipped": (target as HTMLInputElement).checked });
+  }
+
+  /**
+   * カスタム技能を作る。
+   *
+   * `createDoc` は dataset をそのまま作成データに載せる汎用の口だが、技能は名前と
+   * 参照能力値が決まっていないと行を描けないので、先にダイアログで尋ねる。
+   * 「尋ねるかどうか」はプレゼンテーションの決定なので applications 側に置く。
+   */
+  static async _createSkill(this: EmokloreActorSheet, event: Event, _target: HTMLElement) {
+    event.preventDefault();
+
+    const input = await promptCreateSkill();
+    if (!input) return;
+
+    // defaultName / create は ClientDocumentMixin 由来で本体の型に出ないため、
+    // utils/sheet.ts の口を通す（createDoc と同じ経路）
+    const docCls = resolveEmbeddedDocumentClass("Item");
+
+    await docCls.create(
+      {
+        // 名前は空でも通す。あとから鉛筆で直せるので、入力し直しを強いるより
+        // 既定の名前で作ってしまうほうが早い（本体の createDoc と同じ扱い）
+        name: input.name || docCls.defaultName({ type: "skill", parent: this.actor }),
+        type: "skill",
+        system: {
+          category: input.category,
+          characteristicOptions: input.characteristicOptions,
+          // 選べるものが1つでも、判定に使う能力値は明示しておく
+          characteristic: input.characteristicOptions[0],
+          group: input.group,
+        },
+      },
+      { parent: this.actor },
+    );
+  }
+
+  /**
+   * 段入力で、いま選ばれている段をもう一度押したときに値を戻す。
+   *
+   * ラジオは押しても外れないので、0（未修得）に戻す手段がこれしかない。
+   * 段を1つ増やして0を置く手もあるが、バーの左端が常に空いて見えるのでやめた。
+   *
+   * 選択中でない段を押したときは何もしない。ラジオの既定の動作と
+   * submitOnChange に任せる。
+   */
+  static async _selectSegment(this: EmokloreActorSheet, event: Event, target: HTMLElement) {
+    const input = target as HTMLInputElement;
+    const value = Number(input.value);
+    if (!Number.isFinite(value)) return;
+
+    // Number("") は NaN ではなく 0 なので、空文字は「属性が無い」と同じに倒す
+    const raw = input.dataset.clearTo;
+    const clearTo = raw ? Number(raw) : undefined;
+
+    // カスタム技能のレベルはアイテム側が正。段の name はアクター側のミラー
+    // （保存しない枠）を指しているので、フォームの送信に任せると値がどこにも残らない。
+    // 組込技能・能力値は name がそのまま保存先なので、書き込みはフォームに任せる
+    const itemId = input.closest<HTMLElement>("[data-item-id]")?.dataset.itemId;
+    const item = itemId ? this.actor.items.get(itemId) : undefined;
+
+    if (item?.isSkill()) {
+      const next = resolveSegmentValue(value, item.system.level, clearTo);
+      if (next === null || next === item.system.level) return;
+
+      // ラジオの既定動作を止めないと、checked が立ったままアクターのフォームが送られる
+      event.preventDefault();
+      await item.update({ "system.level": next });
+      return;
+    }
+
+    const current = Number(foundry.utils.getProperty(this.actor, input.name));
+    const next = resolveSegmentValue(value, current, clearTo);
+    // 選択中でない段（next === value）はラジオの既定動作と submitOnChange に任せる。
+    // 戻せない入力（能力値は1未満にならない）で押し直したときは null が返る
+    if (next === null || next === value) return;
+
+    // ラジオの既定動作を止めないと、checked が立って submitOnChange が
+    // 元の値で送られ、こちらの更新を打ち消してしまう
+    event.preventDefault();
+    await this.actor.update({ [input.name]: next });
+  }
+
+  /**
+   * カスタム技能の参照能力値を書く。
+   *
+   * 本体の actions はクリックしか見ないので、select の change はフォームの change を
+   * 拾う本体の口（`_onChangeForm`）で受ける。リスナは初回描画で1本張られたきり
+   * 差し替わらないので、描画のたびに繋ぎ直す必要がない。
+   *
+   * 拾ったぶんは super に渡さない。この select は name を持たずアイテム側が保存先なので、
+   * アクターのフォームを送っても何も起きない。
+   */
+  override _onChangeForm(formConfig: unknown, event: Event): void {
+    const select = (event.target as HTMLElement | null)?.closest?.<HTMLSelectElement>(
+      "select[data-skill-characteristic]",
+    );
+
+    if (select) {
+      const itemId = select.closest<HTMLElement>("[data-item-id]")?.dataset.itemId;
+      const item = itemId ? this.actor.items.get(itemId) : undefined;
+      if (item?.isSkill()) void item.update({ "system.characteristic": select.value });
+      return;
+    }
+
+    super._onChangeForm(formConfig, event);
+  }
+
+  /**
+   * タブに属するパートへ自分のタブ情報を渡す（`class="tab"` の active 付けに使う）。
+   *
+   * `context.tabs` は `TABS` を1グループ宣言したシートで本体の `_prepareContext` が積む。
+   * タブを持たないシートでは無いので何もしない。
+   */
+  override async _preparePartContext(
+    partId: string,
+    context: ApplicationRenderContext & { tabs?: Record<string, ApplicationTab>; tab?: unknown },
+    options: EmokloreRenderOptions,
+  ): Promise<ApplicationRenderContext> {
+    await super._preparePartContext(partId, context, options);
+    if (context.tabs && partId in context.tabs) context.tab = context.tabs[partId];
+    return context;
   }
 
   /**
